Print da pilha e de valores de registradores Correção nos operadores em cima de registradores, para o destino ser operado pela origem (I.E. destino dividido pela origem, destino subtraído pela origem, etc)

+++ vm/include/runtime.h
@@ -25,7 +25,7 @@ struct _Runtime
     uint32_t cp;      // Call pointer
     uint32_t sp;      // Stack pointer
     uint32_t ip;      // Instruction pointer
-    uint8_t exit;
+    uint32_t exit;
     char* message;
     bool running;
 };
